refactor: Reorganize platform handling functions
