+++ src/qibolab/_core/dummy/platform.py
@@ -2,8 +2,7 @@ import pathlib
 
 from qibolab._core.components import AcquisitionChannel, DcChannel, IqChannel
 from qibolab._core.instruments.dummy import DummyInstrument, DummyLocalOscillator
-from qibolab._core.parameters import Hardware
-from qibolab._core.platform import Platform
+from qibolab._core.platform import Hardware, Platform
 from qibolab._core.qubits import Qubit
 
 FOLDER = pathlib.Path(__file__).parent
