+++ src/qibolab/_core/instruments/__init__.py
@@ -1,0 +1,5 @@
+from . import abstract
+from .abstract import *
+
+__all__ = []
+__all__ += abstract.__all__
