+++ src/qibolab/_core/platform/__init__.py
@@ -1,7 +1,11 @@
-from . import load, platform
+from . import components, load, parameters, platform
+from .components import *
 from .load import *
+from .parameters import *
 from .platform import *
 
 __all__ = []
+__all__ += components.__all__
 __all__ += load.__all__
+__all__ += parameters.__all__
 __all__ += platform.__all__
