+++ src/qibolab/_core/qubits.py
@@ -1,3 +1,4 @@
+from collections.abc import Mapping
 from typing import Annotated, Any
 
 from pydantic import ConfigDict, Field
@@ -5,7 +6,7 @@ from pydantic import ConfigDict, Field
 from .identifier import ChannelId, QubitId, TransitionId
 from .serialize import Model
 
-__all__ = ["Qubit"]
+__all__ = ["Qubit", "QubitMap"]
 
 DefaultChannelType = Annotated[ChannelId | None, True]
 """If ``True`` the channel is included in the default qubit constructor."""
@@ -75,3 +76,6 @@ class QubitPair(Model):
 
     drive: ChannelId | None = None
     """Output channel, for cross-resonance driving."""
+
+
+QubitMap = Mapping[QubitId, Qubit]
