+++ src/qibolab/platform.py
@@ -1,18 +1,19 @@
-from qibolab._core.parameters import initialize_parameters
+from qibolab._core.platform.components import Hardware
 from qibolab._core.platform.load import (
     PLATFORM,
     PLATFORMS_PATH,
     create_platform,
     load_hardware,
     locate_platform,
-    reset_parameters,
 )
+from qibolab._core.platform.parameters import initialize_parameters, reset_parameters
 from qibolab._core.platform.platform import Platform
 
 __all__ = [
     "PLATFORM",
     "PLATFORMS_PATH",
     "Platform",
+    "Hardware",
     "create_platform",
     "initialize_parameters",
     "load_hardware",
